# apps/kajovo-hotel-api/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceChallengeIn": {
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Device Id",
            "type": "string"
          }
        },
        "required": [
          "device_id"
        ],
        "title": "DeviceChallengeIn",
        "type": "object"
      },
      "DeviceChallengeOut": {
        "properties": {
          "challenge": {
            "title": "Challenge",
            "type": "string"
          },
          "issuedAt": {
            "title": "Issuedat",
            "type": "string"
          },
          "nonce": {
            "title": "Nonce",
            "type": "string"
          }
        },
        "required": [
          "nonce",
          "issuedAt",
          "challenge"
        ],
        "title": "DeviceChallengeOut",
        "type": "object"
      },
      "DeviceRegisterIn": {
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Device Id",
            "type": "string"
          },
          "device_info": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Device Info"
          },
          "device_public_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Public Key"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "required": [
          "device_id"
        ],
        "title": "DeviceRegisterIn",
        "type": "object"
      },
      "DeviceStatusOut": {
        "properties": {
          "activatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activatedat"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DeviceStatusOut",
        "type": "object"
      },
      "DeviceVerifyIn": {
        "properties": {
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "nonce": {
            "minLength": 8,
            "title": "Nonce",
            "type": "string"
          },
          "signature": {
            "minLength": 8,
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "nonce",
          "signature"
        ],
        "title": "DeviceVerifyIn",
        "type": "object"
      },
      "DeviceVerifyOut": {
        "properties": {
          "deviceToken": {
            "title": "Devicetoken",
            "type": "string"
          },
          "device_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Token"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "deviceToken",
          "status"
        ],
        "title": "DeviceVerifyOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "K\u00e1jovoHotel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/device/challenge": {
      "post": {
        "operationId": "device_challenge_device_challenge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceChallengeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceChallengeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device Challenge",
        "tags": [
          "device"
        ]
      }
    },
    "/device/register": {
      "post": {
        "operationId": "device_register_device_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device Register",
        "tags": [
          "device"
        ]
      }
    },
    "/device/status": {
      "get": {
        "operationId": "device_status_device_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device Status",
        "tags": [
          "device"
        ]
      }
    },
    "/device/verify": {
      "post": {
        "operationId": "device_verify_device_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Device-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceVerifyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device Verify",
        "tags": [
          "device"
        ]
      }
    },
    "/device/{device_id}/status": {
      "get": {
        "operationId": "device_status_compat_device__device_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device Status Compat",
        "tags": [
          "device"
        ]
      }
    },
...
  }
}